refactor(rulers): trackrulertype -> trackrulertypeattachment
