+++ src/au3wrap/CMakeLists.txt
@@ -52,8 +52,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/internal/projectsnap.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackcolor.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackcolor.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/trackrulertype.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/trackrulertype.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/trackrulertypeattachment.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/trackrulertypeattachment.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackviewtype.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackviewtype.h
     )
