+++ src/au3wrap/internal/trackrulertypeattachment.h
